multi_agent/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
streamlit>=1.28.0
python-dotenv>=1.0.0
